Clean up/Fixed Formatting issues

+++ src/com/mjr/commands/defaultCommands/AccountLifeCommand.java
@@ -49,7 +49,7 @@ public class AccountLifeCommand extends Command {
 
 		    Date date = new Date(currentTime);
 
-		    long diff = (long) (date.getTime() - parse.getTime());
+		    long diff = date.getTime() - parse.getTime();
 		    long diffDay = diff / (24 * 60 * 60 * 1000);
 		    diff = diff - (diffDay * 24 * 60 * 60 * 1000);
 		    long diffHours = diff / (60 * 60 * 1000);
@@ -62,7 +62,7 @@ public class AccountLifeCommand extends Command {
 		    int diffMonths = (int) (diffDay / 31);
 		    diffDay = diffDay - (diffMonths * 31);
 
-		    int diffYears = (int) (diffMonths / 12);
+		    int diffYears = diffMonths / 12;
 		    diffMonths = diffMonths - (diffYears * 12);
 
 		    Utilities.sendMessage(type, channel, sender + " your twitch account is " + diffYears + " year(s) " + diffMonths
